Conflicts:
	include/SFML/Window/Window.hpp
	src/SFML/Window/Unix/WindowImplX11.cpp
	src/SFML/Window/Win32/WindowImplWin32.cpp
	src/SFML/Window/Win32/WindowImplWin32.hpp
	src/SFML/Window/Window.cpp
	src/SFML/Window/WindowImpl.hpp

+++ include/SFML/Graphics/Font.hpp
@@ -49,7 +49,7 @@ class InputStream;
 ////////////////////////////////////////////////////////////
 class SFML_GRAPHICS_API Font
 {
-public :
+public:
 
     ////////////////////////////////////////////////////////////
     /// \brief Holds various information about a font
@@ -60,7 +60,7 @@ public :
         std::string family; ///< The font family
     };
 
-public :
+public:
 
     ////////////////////////////////////////////////////////////
     /// \brief Default constructor
@@ -181,7 +181,7 @@ public :
     /// \return Kerning value for \a first and \a second, in pixels
     ///
     ////////////////////////////////////////////////////////////
-    int getKerning(Uint32 first, Uint32 second, unsigned int characterSize) const;
+    float getKerning(Uint32 first, Uint32 second, unsigned int characterSize) const;
 
     ////////////////////////////////////////////////////////////
     /// \brief Get the line spacing
@@ -194,7 +194,7 @@ public :
     /// \return Line spacing, in pixels
     ///
     ////////////////////////////////////////////////////////////
-    int getLineSpacing(unsigned int characterSize) const;
+    float getLineSpacing(unsigned int characterSize) const;
 
     ////////////////////////////////////////////////////////////
     /// \brief Get the position of the underline
@@ -209,7 +209,7 @@ public :
     /// \see getUnderlineThickness
     ///
     ////////////////////////////////////////////////////////////
-    int getUnderlinePosition(unsigned int characterSize) const;
+    float getUnderlinePosition(unsigned int characterSize) const;
 
     ////////////////////////////////////////////////////////////
     /// \brief Get the thickness of the underline
@@ -223,7 +223,7 @@ public :
     /// \see getUnderlinePosition
     ///
     ////////////////////////////////////////////////////////////
-    int getUnderlineThickness(unsigned int characterSize) const;
+    float getUnderlineThickness(unsigned int characterSize) const;
 
     ////////////////////////////////////////////////////////////
     /// \brief Retrieve the texture containing the loaded glyphs of a certain size
@@ -249,7 +249,7 @@ public :
     ////////////////////////////////////////////////////////////
     Font& operator =(const Font& right);
 
-private :
+private:
 
     ////////////////////////////////////////////////////////////
     /// \brief Structure defining a row of glyphs
